requirements.txt
yacs
cityscapesScripts
clint
torch
torchvision
transformers
numpy
opencv-python
Pillow
tqdm
anytree
pycocotools
timm
einops
PyYAML
cython
ninja 
